Use nametuples for both received and sent log entries.
